Neatened up files and deleted some, added comments
