+++ slowker/poker_game.py
@@ -1,0 +1,42 @@
+from card_pack import card_pack
+
+
+class poker_game:
+    def __init__(self):
+        self.player_list = []
+        self.board = []
+        return
+    
+    def add_player(self, player):
+        self.player_list.append(player)
+        return
+    
+    def form_player_hands(self):    #The meat of this class - gens, shuffles and deals the cards
+        self.card_pack = card_pack()
+        self.card_pack.gen_card_pack()
+        self.card_pack.shuffle_card_pack()
+        for i in range(len(self.player_list)):
+            for j in range(2):
+                self.player_list[i].hand.append(self.card_pack.pop())
+        return
+    
+    def update_all_hands(self):     #Calculates hand of each player
+        for i in range(len(self.player_list)):
+            self.player_list[i].hand_calculate(self.board)
+        return
+
+    def deal_flop(self):
+        for i in range(3):
+            self.card_pack.pop()
+            self.board.append(self.card_pack.pop())
+        return
+
+    def deal_turn(self):
+        self.card_pack.pop()
+        self.board.append(self.card_pack.pop())
+        return
+
+    def deal_river(self):
+        self.card_pack.pop()
+        self.board.append(self.card_pack.pop())
+        return
